requirements.txt
python==3.9
django==3.1
django-environ==0.4.5
mysqlclient==2.0.3
djangorestframework==3.12.2
dj-rest-auth==2.1.3